fix: stop trapping on out-of-range keycodes from synthetic events
A CGEvent carries its keycode as Int64 and both key-translation entry
points took UInt16, so every one of the nine call sites wrapped it in
UInt16(...) — a trapping conversion. Any process holding Accessibility
can post an event with a keycode outside 0...65535, and this runs on the
tap thread for every keystroke, so a bad value took the app down
mid-⌘Tab rather than being ignored.

Widen translate(keyCode:) and KeyboardLayout.character(for:) to accept
the value as it arrives and narrow once inside with UInt16(exactly:),
returning nil when it doesn't fit. The nine wrapping call sites go away
with it, so no future site can reintroduce the trap.

+++ BetterCmdTab/Switcher/SwitcherController.swift
@@ -2352,11 +2352,11 @@ final class SwitcherController: SwitcherViewDelegate {
         case .letterInputKey(let keyCode):
             // Secure-input Carbon path: resolve the keycode for the current
             // layout, matching the tap's plain letter-jump (lowercased).
-            if let ch = KeyboardLayout.character(for: UInt16(keyCode)) {
+            if let ch = KeyboardLayout.character(for: keyCode) {
                 handleLetter(Character(ch.lowercased()))
             }
         case .searchInputKey(let keyCode):
-            if let ch = KeyboardLayout.character(for: UInt16(keyCode)) {
+            if let ch = KeyboardLayout.character(for: keyCode) {
                 handleSearchInput(ch)
             }
         }
